+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Support for [phpseclib/phpseclib](https://phpseclib.com/) version **3**. #260
 * Support client_secret on token endpoint with PKCE. #293
 
+### Changed
+
+* Allow serializing `OpenIDConnectClient` using `serialize()`
+
 ## [0.9.5]
 
 ### Changed